refactor : k_medi_guide_ai_model.py 파일 삭제 및 모듈, 서비스에 따라 계층 분리
