+++ services/session_store.py
@@ -1,0 +1,7 @@
+# 세션별 상태를 공유하기 위한 글로벌 변수 저장소
+
+session_lang = {}
+session_symptoms = {}
+session_results = {}
+retry_count = {}
+attempts = {}
